+++ internal/nft/diff.go
@@ -7,7 +7,6 @@ import (
 	"os/exec"
 	"regexp"
 	"strings"
-	"time"
 
 	"github.com/adrian-griffin/nfty/internal/colour"
 	"github.com/adrian-griffin/nfty/internal/config"
@@ -130,19 +129,8 @@ func RunDiff(args []string) {
 
 	configPath := fs.Arg(0)
 
-	// hostname for the header line
-	hostname, _ := os.Hostname()
-	now := time.Now().Format("2006-01-02 15:04:05")
-
-	fmt.Printf("  %s %s%s%s\n",
-		colour.Grey("nfty"),
-		colour.Bold("diff"),
-		strings.Repeat(" ", 15),
-		colour.DarkGrey(hostname+" · "+now),
-	)
-	fmt.Println()
-
-	tools.Divider()
+	// print header output
+	tools.CommandExecuteHeader("diff")
 
 	// load and generate the proposed config
 	cfg, err := config.Load(configPath)